hiking/backend-hiking/requirements.txt
blinker==1.6.2
cachelib==0.10.2
click==8.1.3
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
distlib==0.3.6
filelock==3.9.0
Flask==2.3.1
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SQLAlchemy==3.0.3
fonttools==4.38.0
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.6.2
numpy==1.23.5
packaging==21.3
pandas==1.5.1
passlib==1.7.4
Pillow==9.3.0
platformdirs==3.0.0
psycopg2-binary==2.9.6
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2022.6
six==1.16.0
sklearn==0.0.post1
SQLAlchemy==2.0.11
typing_extensions==4.5.0
virtualenv==20.19.0
Werkzeug==2.3.1
